+++ packages/terra-core-docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added
   * Added visual label and associated with input field using 'for' and 'id'.
+  * Added 'ariaLabel' attribute for `terra-form-select` examples.
 
 ## 1.50.0 - (November 13, 2023)